tests/example-contracts/txn-arg/pyproject.toml
[tool.poetry]
name = "txn-arg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "txn_arg"}]

[tool.poetry.dependencies]
python = "^3.11"
beaker-pyteal = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
